refactor(web): P2 — api package swap + carve wave 1 (llm_admin, security, observability)
RFC-003 P2. The protected swap first: src/web/api.py becomes
src/web/api/__init__.py verbatim (relative imports depth-bumped, module
AND indented late imports — the P1 import-surface pin guarded the whole
move). Then 25 sections carve into three domain modules as per-SECTION
registrars called at each section's ORIGINAL position, so the route
registration order the parity contract pins is preserved exactly:

- llm_admin.py (875 ln): pools, codex OAuth, provider mgmt/config,
  ollama, kimi
- security.py: auth, RBAC, host access, API tokens
- observability.py: tools meta, bulkheads, usage, aggregates, audit log,
  log search, risk classification, recovery/branch/validation/
  compression/routing stats, affordances

Cross-section closure state the knife severed, rehomed (declared
deviations from verbatim):
- admin_gate(bot) factory in api_common — ONE canonical
  _auth_configured/_require_admin implementation; the composition root
  and the tokens registrar both build their gate from it
- _codex_creds_lock: module-level in api_common (one lock per process,
  same semantics as the old one-per-table closure)
- the provider-config cluster (_validate_ollama_url, _safe_secret,
  _persist_llm_sections_sync/_persist_config) hoisted to llm_admin
  module scope; the persist pair now takes bot explicitly (6 call sites)

One incident worth reading: pass 2 of the carve script swallowed the
already-inserted register_aggregates call into the moved usage_cost
block — routes still registered (from inside the wrong registrar!) so
the parity COUNT stayed 183 and tests stayed green; an import-vs-call
audit caught it. The call is restored at its original position and
parity's ORDER check now proves the fix.

Lint: the file rename made every pre-existing finding in api.py 'new'
to the gate, which pulls the P6 burn-down forward for these files —
all of it fixed here: gate reports new=0, resolved=58. Suite 6,079.
